+++ R/paths.R
@@ -122,17 +122,10 @@ renv_paths_root_default <- function() {
     !is.na(Sys.getenv("TESTTHAT", unset = NA))
 
   if (checking)
-    return(renv_paths_root_default_impl())
+    return(renv_paths_root_default_tempdir())
 
-  root <- switch(
-    Sys.info()[["sysname"]],
-    Darwin  = Sys.getenv("XDG_DATA_HOME", "~/Library/Application Support"),
-    Windows = Sys.getenv("LOCALAPPDATA", Sys.getenv("APPDATA")),
-    Sys.getenv("XDG_DATA_HOME", "~/.local/share")
-  )
-
-  root <- renv_path_normalize(root, winslash = "/", mustWork = FALSE)
-  path <- file.path(root, "renv")
+  # resolve path to cache
+  path <- renv_paths_root_default_impl()
 
   # check for user consent
   consenting <- identical(getOption("renv.consenting"), TRUE)
@@ -165,19 +158,70 @@ renv_paths_root_default <- function() {
     )
   }
 
-  renv_paths_root_default_impl()
+  renv_paths_root_default_tempdir()
 
 }
 
 renv_paths_root_default_impl <- function() {
+
+  # support renv projects created with the old root location
+  oldroot <- switch(
+    Sys.info()[["sysname"]],
+    Darwin  = Sys.getenv("XDG_DATA_HOME", "~/Library/Application Support"),
+    Windows = Sys.getenv("LOCALAPPDATA", Sys.getenv("APPDATA")),
+    Sys.getenv("XDG_DATA_HOME", "~/.local/share")
+  )
+
+  # if we've already initialized renv with the old location, use it
+  oldpath <- file.path(oldroot, "renv")
+  if (file.exists(oldpath))
+    return(oldpath)
+
+  # try using tools to get the user directory
+  tools <- asNamespace("tools")
+  path <- catch(tools$R_user_dir("renv", which = "cache"))
+  if (!inherits(path, "error"))
+    return(path)
+
+  # try using our own backfill for older versions of R
+  renv_paths_root_default_impl_fallback()
+
+}
+
+renv_paths_root_default_impl_fallback <- function() {
+
+  # check for R_USER_CACHE_DIR + XDG_CACHE_HOME overrides
+  envvars <- c("R_USER_CACHE_DIR", "XDG_CACHE_HOME")
+  for (envvar in envvars) {
+    root <- Sys.getenv(envvar, unset = NA)
+    if (!is.na(root)) {
+      path <- file.path(root, "R/renv")
+      return(path)
+    }
+  }
+
+  # use platform-specific default fallbacks
+  if (renv_platform_windows())
+    file.path(Sys.getenv("LOCALAPPDATA"), "R/cache/R/renv")
+  else if (renv_platform_macos())
+    "~/Library/Caches/org.R-project.R/R/renv"
+  else
+    "~/.cache/R/renv"
+
+}
+
+renv_paths_root_default_tempdir <- function() {
   temp <- file.path(tempdir(), "renv")
   ensure_directory(temp)
   return(temp)
 }
-
 # nocov end
 
 renv_paths_init <- function() {
+  renv_paths_init_envvars()
+}
+
+renv_paths_init_envvars <- function() {
 
   envvars <- Sys.getenv()
 
@@ -205,7 +249,20 @@ renv_paths_init <- function() {
 #' Windows      \tab `%LOCALAPPDATA%/renv` \cr
 #' }
 #'
-#' If desired, this path can be adjusted by setting the `RENV_PATHS_ROOT`
+#' For new installations of `renv` using R (>= 4.0.0), `renv` will use
+#' [tools::R_user_dir()] to resolve the root directory. If an `renv` root
+#' directory has already been created in one of the old locations, that will
+#' still be used. This change was made to comply with the CRAN policy
+#' requirements of \R packages. By default, these paths resolve as:
+#'
+#' \tabular{ll}{
+#' **Platform** \tab **Location** \cr
+#' Linux        \tab `~/.cache/R/renv` \cr
+#' macOS        \tab `~/Library/Caches/org.R-project.R/R/renv` \cr
+#' Windows      \tab `%LOCALAPPDATA%/R/cache/R/renv` \cr
+#' }
+#'
+#' If desired, this path can be customized by setting the `RENV_PATHS_ROOT`
 #' environment variable. This can be useful if you'd like, for example, multiple
 #' users to be able to share a single global cache.
 #'