+++ tests/testthat/test-paths.R
@@ -72,3 +72,12 @@ test_that("UTF-8 paths can be normalized", {
   expect_equal(expected, actual)
 
 })
+
+test_that("our fallback R_user_dir() implementation is compatible", {
+
+  skip_if(getRversion() < "4.0.0")
+  actual   <- renv_paths_root_default_impl_fallback()
+  expected <- tools::R_user_dir("renv", which = "cache")
+  expect_true(renv_path_same(actual, expected))
+
+})